+++ src/Workspaces/Core/Portable/OrganizeImports/OrganizeImportsOptionsProviders.cs
@@ -2,29 +2,13 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Runtime.Serialization;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.CodeAnalysis.AddImport;
 using Microsoft.CodeAnalysis.Editing;
 using Microsoft.CodeAnalysis.Formatting;
 using Microsoft.CodeAnalysis.Options;
 
 namespace Microsoft.CodeAnalysis.OrganizeImports;
-
-[DataContract]
-internal readonly record struct OrganizeImportsOptions
-{
-    [DataMember] public bool PlaceSystemNamespaceFirst { get; init; } = AddImportPlacementOptions.Default.PlaceSystemNamespaceFirst;
-    [DataMember] public bool SeparateImportDirectiveGroups { get; init; } = SyntaxFormattingOptions.CommonDefaults.SeparateImportDirectiveGroups;
-    [DataMember] public string NewLine { get; init; } = LineFormattingOptions.Default.NewLine;
-
-    public OrganizeImportsOptions()
-    {
-    }
-
-    public static readonly OrganizeImportsOptions Default = new();
-}
 
 internal static class OrganizeImportsOptionsProviders
 {
