+++ src/Workspaces/SharedUtilitiesAndExtensions/Workspace/Core/CodeGeneration/CleanCodeGenerationOptions.cs
@@ -8,9 +8,9 @@ using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.CodeCleanup;
 using Microsoft.CodeAnalysis.Diagnostics.Analyzers.NamingStyles;
 
-#if !CODE_STYLE
+//#if !CODE_STYLE
 using Microsoft.CodeAnalysis.Host;
-#endif
+//#endif
 
 namespace Microsoft.CodeAnalysis.CodeGeneration;
 
@@ -23,7 +23,6 @@ internal readonly record struct CleanCodeGenerationOptions
     [DataMember]
     public required CodeCleanupOptions CleanupOptions { get; init; }
 
-#if !CODE_STYLE
     public static CleanCodeGenerationOptions GetDefault(LanguageServices languageServices)
         => new()
         {
@@ -37,7 +36,6 @@ internal readonly record struct CleanCodeGenerationOptions
             GenerationOptions = GenerationOptions,
             AddImportOptions = CleanupOptions.AddImportOptions
         };
-#endif
 }
 
 #if !CODE_STYLE
