+++ src/Workspaces/SharedUtilitiesAndExtensions/Workspace/Core/CodeGeneration/CodeGenerationOptions.cs
@@ -3,16 +3,11 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Runtime.Serialization;
-using System.Threading;
-using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.AddImport;
-using Roslyn.Utilities;
-using Microsoft.CodeAnalysis.Diagnostics.Analyzers.NamingStyles;
 using Microsoft.CodeAnalysis.CodeStyle;
-using Microsoft.CodeAnalysis.Options;
-using Microsoft.CodeAnalysis.CodeCleanup;
-
+using Microsoft.CodeAnalysis.Diagnostics.Analyzers.NamingStyles;
 using Microsoft.CodeAnalysis.Host;
+using Microsoft.CodeAnalysis.Options;
 //#if !CODE_STYLE
 //using Microsoft.CodeAnalysis.Host;
 //#endif
@@ -40,10 +35,8 @@ internal record CodeGenerationOptions
         NamingStyle = options.GetOption(NamingStyleOptions.NamingPreferences, language);
     }
 
-#if !CODE_STYLE
     public static CodeGenerationOptions GetDefault(LanguageServices languageServices)
         => languageServices.GetRequiredService<ICodeGenerationService>().DefaultOptions;
-#endif
 }
 
 [DataContract]
@@ -55,12 +48,10 @@ internal readonly record struct CodeAndImportGenerationOptions
     [DataMember]
     public required AddImportPlacementOptions AddImportOptions { get; init; }
 
-#if !CODE_STYLE
     internal static CodeAndImportGenerationOptions GetDefault(LanguageServices languageServices)
         => new()
         {
             GenerationOptions = CodeGenerationOptions.GetDefault(languageServices),
             AddImportOptions = AddImportPlacementOptions.Default
         };
-#endif
 }
